+++ zooFrontend/src/pages/admin/manage-users/UserDetails.jsx
@@ -1,7 +1,7 @@
 import React, { useEffect, useState } from 'react';
-import {Link, useParams} from 'react-router-dom';
+import {useParams} from 'react-router-dom';
 import Navbar from "../Navbar";
-import {getUserWorkSchedule} from '../../../services/AdminService';
+import {getUserWorkSchedule} from '../../../services/adminService';
