+++ zooFrontend/src/routes/HomeRedirect.jsx
@@ -1,5 +1,5 @@
 import { Navigate } from 'react-router-dom';
-import { useAuth } from '../services/AuthContext';
+import { useAuth } from '../context/AuthContext';
 
 const HomeRedirect = () => {
     const { user } = useAuth();
