+++ zooFrontend/src/routes/PrivateRoute.jsx
@@ -1,5 +1,5 @@
 import { Navigate, Outlet } from 'react-router-dom';
-import { useAuth } from '../services/AuthContext';
+import { useAuth } from '../context/AuthContext';
 
 const PrivateRoute = ({ requiredRole }) => {
     const { user, authChecked } = useAuth();
