+++ zooFrontend/src/services/adminService.js
@@ -1,14 +1,8 @@
 import axiosInstance from "./axiosInstance";
 
 export const getAllUsers = async () => {
-    const token = localStorage.getItem('accessToken');
-
     try {
-        const response = await axiosInstance.get(`api/admin/users`, {
-            headers: {
-                Authorization: `Bearer ${token}`,
-            },
-        });
+        const response = await axiosInstance.get(`api/admin/users`);
         return response.data;
     } catch (error) {
         console.error('Błąd podczas pobierania użytkowników:', error);
@@ -17,14 +11,8 @@ export const getAllUsers = async () => {
 };
 
 export const getUserWorkSchedule = async (id) => {
-    const token = localStorage.getItem('accessToken');
-
     try {
-        const response = await axiosInstance.get(`api/admin/user/${id}/work-schedule`, {
-            headers: {
-                Authorization: `Bearer ${token}`,
-            },
-        });
+        const response = await axiosInstance.get(`api/admin/user/${id}/work-schedule`);
         return response.data;
     } catch (error) {
         console.error('Błąd podczas pobierania harmonogramu użytkownika:', error);
